refactor: split runner and runnerPool tests
