+++ src/consumer/__tests__/runnerPool.spec.js
@@ -1,4 +1,3 @@
-const Runner = require('../runner')
 const Batch = require('../batch')
 const {
   KafkaJSProtocolError,
@@ -17,17 +16,14 @@ const REBALANCE_IN_PROGRESS = 27
 const rebalancingError = () => new KafkaJSProtocolError(createErrorFromCode(REBALANCE_IN_PROGRESS))
 
 describe('Consumer > RunnerPool', () => {
-  let runner,
-    /** @type {ReturnType<typeof createRunnerPool>} */
-    runnerPool,
+  let runnerPool,
     consumerGroup,
     onCrash,
     eachBatch,
     topicName,
     partition,
     emptyBatch,
-    instrumentationEmitter,
-    recover
+    instrumentationEmitter
 
   beforeEach(() => {
     topicName = `topic-${secureRandom()}`
@@ -41,7 +37,6 @@ describe('Consumer > RunnerPool', () => {
 
     eachBatch = jest.fn()
     onCrash = jest.fn()
-    recover = jest.fn()
     consumerGroup = {
       connect: jest.fn(),
       join: jest.fn(),
@@ -58,12 +53,6 @@ describe('Consumer > RunnerPool', () => {
       isLeader: jest.fn(() => true),
     }
     instrumentationEmitter = new InstrumentationEventEmitter()
-    runner = new Runner({
-      consumerGroup,
-      instrumentationEmitter,
-      logger: newLogger(),
-      eachBatch,
-    })
     runnerPool = createRunnerPool({
       consumerGroup,
       instrumentationEmitter,
@@ -74,7 +63,6 @@ describe('Consumer > RunnerPool', () => {
   })
 
   afterEach(async () => {
-    runner && (await runner.stop())
     runnerPool && (await runnerPool.stop())
   })
 
@@ -89,136 +77,7 @@ describe('Consumer > RunnerPool', () => {
         })
         .mockImplementationOnce(() => true)
 
-      runner.scheduleConsume = jest.fn()
-      await runner.start({ recover })
-      expect(runner.scheduleConsume).toHaveBeenCalled()
-      expect(onCrash).not.toHaveBeenCalled()
-    })
-  })
-
-  it('should "commit" offsets during fetch', async () => {
-    const batch = new Batch(topicName, 0, {
-      partition,
-      highWatermark: 5,
-      messages: [{ offset: 4, key: '1', value: '2' }],
-    })
-
-    consumerGroup.nextBatch.mockImplementationOnce((_, callback) => callback(batch))
-    runner.scheduleConsume = jest.fn()
-    await runner.start({ recover })
-    await runner.consume() // Manually fetch for test
-    expect(eachBatch).toHaveBeenCalled()
-    expect(consumerGroup.commitOffsets).toHaveBeenCalled()
-    expect(onCrash).not.toHaveBeenCalled()
-  })
-
-  describe('"eachBatch" callback', () => {
-    it('allows providing offsets to "commitOffsetIfNecessary"', async () => {
-      const batch = new Batch(topicName, 0, {
-        partition,
-        highWatermark: 5,
-        messages: [{ offset: 4, key: '1', value: '2' }],
-      })
-
-      consumerGroup.nextBatch.mockImplementationOnce((_, callback) => callback(batch))
-      runner.scheduleConsume = jest.fn()
-      await runner.start({ recover })
-      await runner.consume() // Manually fetch for test
-
-      expect(eachBatch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          commitOffsetsIfNecessary: expect.any(Function),
-        })
-      )
-
-      const { commitOffsetsIfNecessary } = eachBatch.mock.calls[0][0] // Access the callback
-
-      // Clear state
-      consumerGroup.commitOffsetsIfNecessary.mockClear()
-      consumerGroup.commitOffsets.mockClear()
-
-      // No offsets provided
-      await commitOffsetsIfNecessary()
-      expect(consumerGroup.commitOffsetsIfNecessary).toHaveBeenCalledTimes(1)
-      expect(consumerGroup.commitOffsets).toHaveBeenCalledTimes(0)
-
-      // Clear state
-      consumerGroup.commitOffsetsIfNecessary.mockClear()
-      consumerGroup.commitOffsets.mockClear()
-
-      // Provide offsets
-      const offsets = {
-        topics: [{ topic: topicName, partitions: [{ offset: '1', partition: 0 }] }],
-      }
-
-      await commitOffsetsIfNecessary(offsets)
-      expect(consumerGroup.commitOffsetsIfNecessary).toHaveBeenCalledTimes(0)
-      expect(consumerGroup.commitOffsets).toHaveBeenCalledTimes(1)
-      expect(consumerGroup.commitOffsets).toHaveBeenCalledWith(offsets)
-    })
-  })
-
-  describe('when eachBatchAutoResolve is set to false', () => {
-    beforeEach(() => {
-      runner = new Runner({
-        consumerGroup,
-        instrumentationEmitter: new InstrumentationEventEmitter(),
-        eachBatchAutoResolve: false,
-        eachBatch,
-        logger: newLogger(),
-      })
-      runner.scheduleConsume = jest.fn(() => runner.consume())
-    })
-
-    it('does not call resolveOffset with the last offset', async () => {
-      const batch = new Batch(topicName, 0, {
-        partition,
-        highWatermark: 5,
-        messages: [{ offset: 4, key: '1', value: '2' }],
-      })
-
-      consumerGroup.nextBatch.mockImplementationOnce((_, callback) => callback(batch))
-      await runner.start({ recover })
-      expect(onCrash).not.toHaveBeenCalled()
-      expect(consumerGroup.resolveOffset).not.toHaveBeenCalled()
-    })
-  })
-
-  describe('when autoCommit is set to false', () => {
-    let eachBatchCallUncommittedOffsets
-
-    beforeEach(() => {
-      eachBatchCallUncommittedOffsets = jest.fn(async ({ uncommittedOffsets }) => {
-        uncommittedOffsets()
-      })
-
-      runner = new Runner({
-        consumerGroup,
-        instrumentationEmitter: new InstrumentationEventEmitter(),
-        eachBatch: eachBatchCallUncommittedOffsets,
-        autoCommit: false,
-        logger: newLogger(),
-      })
-      runner.scheduleConsume = jest.fn(() => runner.consume())
-    })
-
-    it('should not commit offsets during fetch', async () => {
-      const batch = new Batch(topicName, 0, {
-        partition,
-        highWatermark: 5,
-        messages: [{ offset: 4, key: '1', value: '2' }],
-      })
-
-      consumerGroup.nextBatch.mockImplementationOnce((_, callback) => callback(batch))
-      runner.scheduleConsume = jest.fn()
-      await runner.start({ recover })
-      await runner.consume() // Manually fetch for test
-
-      expect(consumerGroup.commitOffsets).not.toHaveBeenCalled()
-      expect(consumerGroup.commitOffsetsIfNecessary).not.toHaveBeenCalled()
-      expect(eachBatchCallUncommittedOffsets).toHaveBeenCalled()
-      expect(consumerGroup.uncommittedOffsets).toHaveBeenCalled()
-
+      await runnerPool.start()
       expect(onCrash).not.toHaveBeenCalled()
     })
   })
@@ -314,29 +173,6 @@ describe('Consumer > RunnerPool', () => {
       await expect(onCrash).toHaveBeenCalledWith(expect.any(KafkaJSNumberOfRetriesExceeded))
     })
 
-    it('correctly catch exceptions in parallel "heartbeat" processing', async () => {
-      const batch = new Batch(topicName, 0, {
-        partition,
-        highWatermark: 5,
-        messages: [{ offset: 4, key: '1', value: '2' }],
-      })
-
-      consumerGroup.heartbeat = async () => {
-        throw new Error('Error while processing heartbeats in parallel')
-      }
-
-      consumerGroup.nextBatch
-        .mockImplementationOnce(async (_, callback) => callback(await sleep(100)))
-        .mockImplementationOnce(async (_, callback) => callback(batch))
-
-      runner.scheduleConsume = jest.fn()
-      await runner.start({ recover })
-
-      await expect(runner.consume()).rejects.toThrow(
-        'Error while processing heartbeats in parallel'
-      )
-    })
-
     it('a triggered rejoin failing should cause a crash', async () => {
       await runnerPool.start()
 
@@ -352,22 +188,6 @@ describe('Consumer > RunnerPool', () => {
 
       await waitFor(() => onCrash.mock.calls.length > 0)
       expect(onCrash).toHaveBeenCalledWith(unknownError)
-    })
-
-    it('should ignore request errors from nextBatch on stopped consumer', async () => {
-      consumerGroup.nextBatch
-        .mockImplementationOnce(async () => {
-          await sleep(10)
-          throw new Error('Failed or manually rejected request')
-        })
-        .mockImplementationOnce(async (_, callback) => callback(await sleep(10)))
-
-      runner.scheduleConsume = jest.fn()
-      await runner.start({ recover })
-      runner.running = false
-
-      runner.consume()
-      await sleep(100)
     })
   })
 })
